That's what I get for merging contributions and not testing them

+++ src/com/matburt/mobileorg/Synchronizers/DropboxSynchronizer.java
@@ -117,6 +117,9 @@ public class DropboxSynchronizer extends Synchronizer {
         if (filesInDb.size() > 0) {
             Object[] arrObj = filesInDb.toArray();
             for (int i = 0; i < arrObj.length; i++) {
+                if (((String)arrObj[i]).equals("mobileorg.org")) {
+                    continue;
+                }
                 Log.i(LT, "Orphaned file: " + (String)arrObj[i]);
                 removeFile((String)arrObj[i]);
             }
